(hopefully) fixed data abort exception and antijam
in antijam, the time counter variable was being recreated every time the loop was run, not allowing it to get anywhere. for the screen, pressing the buttons was attempting to access a screen at an index at the list that didn't exist. i replaced it with a boolean that toggles back and forth, which is how it was before i added the auton builder